VSDEV-1762: Remove use of find_endpoints parameter for APIConfig
The find_endpoints parameter was introduced to support the two configuration
methods. As we only have to support one, we can do without this parameter.
Instead we pass the endpoints directly to APIConfig.

This commit also updates the APIConfig tests to take this change into account.
As part of that work, it removes a lot of duplicated code from these tests.